+++ CHANGELOG.md
@@ -1,6 +1,113 @@
-## 0.3.8 (unreleased)
+## 0.4.0 (unreleased)
 
+FEATURES:
 
+* builder/amazon/ebs: Ability to specify which availability zone to create
+  instance in. [GH-536]
+
+BUG FIXES:
+
+* core: Won't panic when writing to a bad pipe. [GH-560]
+* common/uuid: Use cryptographically secure PRNG when generating
+  UUIDs. [GH-552]
+
+## 0.3.10 (October 20, 2013)
+
+FEATURES:
+
+* Ansible provisioner
+
+IMPROVEMENTS:
+
+* post-processor/vagrant: support instance-store AMIs built by Packer. [GH-502]
+* post-processor/vagrant: can now specify compression level to use
+  when creating the box. [GH-506]
+
+BUG FIXES:
+
+* builder/all: timeout waiting for SSH connection is a failure. [GH-491]
+* builder/amazon: Scrub sensitive data from the logs. [GH-521]
+* builder/amazon: Handle the situation where an EC2 instance might not
+  be immediately available. [GH-522]
+* builder/amazon/chroot: Files copied into the chroot remove destination
+  before copy, fixing issues with dangling symlinks. [GH-500]
+* builder/digitalocean: don't panic if erroneous API response doesn't
+  contain error message. [GH-492]
+* builder/digitalocean: scrub API keys from config debug output [GH-516]
+* builder/virtualbox: error if VirtualBox version cant be detected. [GH-488]
+* builder/virtualbox: detect if vboxdrv isn't properly setup. [GH-488]
+* builder/virtualbox: sleep a bit before export to ensure the sesssion
+  is unlocked. [GH-512]
+* builder/virtualbox: create SATA drives properly on VirtualBox 4.3 [GH-547]
+* builder/virtualbox: support user templates in SSH key path. [GH-539]
+* builder/vmware: support user templates in SSH key path. [GH-539]
+* communicator/ssh: Fix issue where a panic could arise from a nil
+  dereference. [GH-525]
+* post-processor/vagrant: Fix issue with VirtualBox OVA. [GH-548]
+* provisioner/salt: Move salt states to correct remote directory. [GH-513]
+* provisioner/shell: Won't block on certain scripts on Windows anymore.
+  [GH-507]
+
+## 0.3.9 (October 2, 2013)
+
+FEATURES:
+
+* The Amazon chroot builder is now able to run without any `sudo` privileges
+  by using the "command_wrapper" configuration. [GH-430]
+* Chef provisioner supports environments. [GH-483]
+
+BUG FIXES:
+
+* core: default user variable values don't need to be strings. [GH-456]
+* builder/amazon-chroot: Fix errors with waitin for state change. [GH-459]
+* builder/digitalocean: Use proper error message JSON key (DO API change).
+* communicator/ssh: SCP uploads now work properly when directories
+  contain symlinks. [GH-449]
+* provisioner/chef-solo: Data bags and roles path are now properly
+  populated when set. [GH-470]
+* provisioner/shell: Windows line endings are actually properly changed
+  to Unix line endings. [GH-477]
+
+## 0.3.8 (September 22, 2013)
+
+FEATURES:
+
+* core: You can now specify `only` and `except` configurations on any
+  provisioner or post-processor to specify a list of builds that they
+  are valid for. [GH-438]
+* builders/virtualbox: Guest additions can be attached rather than uploaded,
+  easier to handle for Windows guests. [GH-405]
+* provisioner/chef-solo: Ability to specify a custom Chef configuration
+  template.
+* provisioner/chef-solo: Roles and data bags support. [GH-348]
+
+IMPROVEMENTS:
+
+* core: User variables can now be used for integer, boolean, etc.
+  values. [GH-418]
+* core: Plugins made with incompatible versions will no longer load.
+* builder/amazon/all: Interrupts work while waiting for AMI to be ready.
+* provisioner/shell: Script line-endings are automatically converted to
+  Unix-style line-endings. Can be disabled by setting "binary" to "true".
+  [GH-277]
+
+BUG FIXES:
+
+* core: Set TCP KeepAlives on internally created RPC connections so that
+  they don't die. [GH-416]
+* builder/amazon/all: While waiting for AMI, will detect "failed" state.
+* builder/amazon/all: Waiting for state will detect if the resource (AMI,
+  instance, etc.) disappears from under it.
+* builder/amazon/instance: Exclude only contents of /tmp, not /tmp
+  itself. [GH-437]
+* builder/amazon/instance: Make AccessKey/SecretKey available to bundle
+  command even when they come from the environment. [GH-434]
+* builder/virtualbox: F1-F12 and delete scancodes now work. [GH-425]
+* post-processor/vagrant: Override configurations properly work. [GH-426]
+* provisioner/puppet-masterless: Fix failure case when both facter vars
+  are used and prevent_sudo. [GH-415]
+* provisioner/puppet-masterless: User variables now work properly in
+  manifest file and hiera path. [GH-448]
 
 ## 0.3.7 (September 9, 2013)